Split ioutil into its own source file
